refactor - api 서비스 폴더 리팩토링 (#79)
* refactor: 서비스명 수정

* refactor: Ai요약 api 유저서비스로 이전 및 남자, 여자 api 구분

* refactor: 모델 파일 통합작업

* refactor: 모델 구조 통합
