+++ app/src/main/java/com/apptive/japkor/LoginCallbackActivity.kt
@@ -123,7 +123,7 @@ class LoginCallbackActivity : ComponentActivity() {
 
     private suspend fun refreshMemberInfo(dataStore: DataStoreManager) {
         runCatching {
-            ServiceFactory.memberService.getMemberInfo().awaitResponse()
+            ServiceFactory.memberService.getUserInfo().awaitResponse()
         }.onSuccess { response ->
             if (response.isSuccessful) {
                 val body = response.body()
