+++ app/src/main/java/com/apptive/japkor/data/api/MatchingService.kt
@@ -1,6 +1,5 @@
 package com.apptive.japkor.data.api
 
-import com.apptive.japkor.data.model.AiSummaryResponse
 import com.apptive.japkor.data.model.MalePendingMatchingResponse
 import com.apptive.japkor.data.model.MatchingResponse
 import retrofit2.Call
@@ -9,17 +8,17 @@ import retrofit2.http.POST
 import retrofit2.http.Path
 
 interface MatchingService {
-    @GET("members/ai-summary")
-    fun getMyAiSummary(): Call<AiSummaryResponse>
 
+    // 여자 전용
     @GET("members/matchings/female")
     fun getFemaleMatchings(): Call<List<MatchingResponse>>
 
-    @GET("members/matchings/male/pendingMatching")
-    fun getMalePendingMatchings(): Call<List<MalePendingMatchingResponse>>
-
     @POST("members/matchings/{matchingId}/select")
     fun femaleSelectMatching(@Path("matchingId") matchingId: Long): Call<Void>
+
+    // 남자 전용
+    @GET("members/matchings/male/pendingMatching")
+    fun getMalePendingMatchings(): Call<List<MalePendingMatchingResponse>>
 
     @POST("members/matchings/{matchingId}/accept")
     fun maleAcceptMatching(@Path("matchingId") matchingId: Long): Call<Void>
