+++ app/src/main/java/com/apptive/japkor/data/api/UserService.kt
@@ -1,0 +1,14 @@
+package com.apptive.japkor.data.api
+
+import com.apptive.japkor.data.model.AiSummaryResponse
+import com.apptive.japkor.data.model.UserInfoResponse
+import retrofit2.Call
+import retrofit2.http.GET
+
+interface UserService {
+    @GET("members/ai-summary")
+    fun getMyAiSummary(): Call<AiSummaryResponse>
+
+    @GET("members/info")
+    fun getUserInfo(): Call<UserInfoResponse>
+}
