+++ app/src/main/java/com/apptive/japkor/data/model/Error.kt
@@ -1,6 +1,5 @@
 package com.apptive.japkor.data.model
 
-
 data class ErrorResponse(
     val message: String?
 )
