--- app/src/main/java/com/apptive/japkor/data/model/MalePendingMatchingResponse.kt
@@ -1,19 +1,0 @@
-package com.apptive.japkor.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class MalePendingMatchingResponse(
-    val matchingId: Long,
-    val femaleMemberId: Long,
-    val femaleName: String,
-    val femaleEmail: String,
-    val height: Int?,
-    val weight: Int?,
-    val residenceArea: String?,
-    val aiSummary: String?,
-    val status: String,
-    val createdAt: String,
-    val thumbnailImageUrl: String?,
-    @SerializedName(value = "profileImageUrls", alternate = ["profileImageUrl"])
-    val profileImageUrls: List<String>?
-)
